requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
spacy==3.8.14
scikit-learn==1.5.0
pymongo==4.7.2
motor==3.4.0
pydantic==2.7.4
python-dotenv==1.0.1